requirements.txt
asgiref==3.11.1
Django==6.0.3
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==40.12.0
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.2
pytest-django==4.12.0
python-decouple==3.8
sqlparse==0.5.5
